test: move mock object to separate file
